Update project#1
Operator's choice on a created synthetic dataset. We try to predict if an operator chooses the manual solution and recovery solver during an airline disruption. The dataset was manipulated, cleaned, and azure was used to create synthetic data on the existing data to make it more unique. Multi-Layer perceptron, Xtreme Gradient Boosting, and Support Vector Machine models were used in predicting the operator's choice and a comparison was made comparing various performance metrics